server/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
anthropic==0.20.0
python-dotenv==1.0.0
pydantic==2.4.2
python-multipart==0.0.6

# Testing dependencies
pytest==7.4.0
httpx==0.24.1
pytest-asyncio==0.21.1
pytest-mock==3.11.1
pytest-cov==4.1.0python-jose[cryptography]
passlib[bcrypt]
python-multipart

# Database
sqlalchemy
psycopg2-binary
alembic
